Finished folding the push methods into analytics, with push-fcm serving as just a basic implementation of FCM class.
