+++ sdk/push-fcm/src/main/java/com/klaviyo/push_fcm/KlaviyoPushService.kt
@@ -1,12 +1,8 @@
 package com.klaviyo.push_fcm
 
-import android.content.Intent
 import com.google.firebase.messaging.FirebaseMessagingService
 import com.google.firebase.messaging.RemoteMessage
 import com.klaviyo.analytics.Klaviyo
-import com.klaviyo.analytics.model.Event
-import com.klaviyo.analytics.model.EventKey
-import com.klaviyo.analytics.model.EventType
 
 /**
  * Implementation of the FCM messaging service that runs when the parent application is started
@@ -16,74 +12,6 @@ import com.klaviyo.analytics.model.EventType
  * that the implementation details of this service are carried over into their own
  */
 class KlaviyoPushService : FirebaseMessagingService() {
-
-    companion object {
-
-        /**
-         * Saves the device FCM push token and registers to the current profile
-         *
-         * We append this token to a property map and queue it into an identify request to send to
-         * the Klaviyo asynchronous APIs.
-         * We then write it into the shared preferences so that we can fetch the token for this
-         * device as needed
-         *
-         * @see FirebaseMessagingService.onNewToken()
-         * @param pushToken The push token provided by the FCM Service
-         */
-        fun setPushToken(pushToken: String) = Klaviyo.setPushToken(pushToken)
-
-        /**
-         * Retrieves the device FCM push token stored on this device
-         *
-         * @return The push token we read from the data store
-         */
-        internal fun getPushToken(): String? = Klaviyo.getPushToken()
-
-        /**
-         * Logs an $opened_push event for a remote notification that originated from Klaviyo
-         *
-         * @param payload The data attributes of the push notification payload
-         */
-        internal fun openedPush(payload: Map<String, String>) {
-            if (!isKlaviyoPush(payload)) return // Track only pushes originating from klaviyo
-
-            val event = Event(
-                EventType.OPENED_PUSH,
-                payload.mapKeys {
-                    EventKey.CUSTOM(it.key)
-                }
-            )
-
-            getPushToken()?.let { event[EventKey.PUSH_TOKEN] = it }
-
-            Klaviyo.createEvent(event)
-        }
-
-        /**
-         * Handles a push received while app is in the foreground
-         *
-         * @param message Message received via [FirebaseMessagingService.onMessageReceived]
-         */
-        fun handlePush(message: RemoteMessage): Boolean {
-            // TODO could we display the notification in system tray?
-            // NOTE this is where we'd handle a data-only push, when we support those
-            return isKlaviyoPush(message.data)
-        }
-
-        /**
-         * Logs an $opened_push event for a notification that originated from Klaviyo
-         * After being opened from the system tray
-         *
-         * @param intent The Intent generated from tapping the notification
-         */
-        fun handlePush(intent: Intent?) = openedPush(
-            intent?.extras?.let { extras ->
-                extras.keySet().associateWith { key -> extras.getString(key, "") }
-            } ?: emptyMap()
-        )
-
-        private fun isKlaviyoPush(data: Map<String, String>): Boolean = data.containsKey("_k")
-    }
 
     /**
      * FCM service calls this function whenever a token is generated
@@ -95,7 +23,7 @@ class KlaviyoPushService : FirebaseMessagingService() {
      */
     override fun onNewToken(newToken: String) {
         super.onNewToken(newToken)
-        setPushToken(newToken)
+        Klaviyo.setPushToken(newToken)
     }
 
     /**
@@ -106,7 +34,7 @@ class KlaviyoPushService : FirebaseMessagingService() {
      * @param message Remote message that has been received
      */
     override fun onMessageReceived(message: RemoteMessage) {
+        // TODO could we provide a way to display the notification in system tray?
         super.onMessageReceived(message)
-        handlePush(message)
     }
 }
